feat: extract search into standalone fragment
- Create SearchFragment with own layout and search logic
- Add search destination to nav graph
- Remove search overlay from MainActivity (views, query flow, adapter)
- Search now navigates via navController, pops itself on film click
