+++ app/src/main/res/layout/activity_main.xml
@@ -37,30 +37,4 @@
         app:headerLayout="@layout/nav_header"
         app:menu="@menu/drawer_nav_menu" />
 
-    <LinearLayout
-        android:id="@+id/search_layout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/transparent"
-        android:gravity="top"
-        android:orientation="vertical"
-        android:visibility="gone">
-
-        <EditText
-            android:id="@+id/search_bar"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:background="@drawable/rounded"
-            android:hint="@string/enter_search_keywords"
-            android:padding="10dp"
-            android:singleLine="true"
-            tools:ignore="Autofill,TextFields" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/search_results"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_margin="10dp" />
-    </LinearLayout>
 </androidx.drawerlayout.widget.DrawerLayout>
